+++ xla/pjrt/c/pjrt_c_api.h
@@ -111,7 +111,7 @@ PJRT_DEFINE_STRUCT_TRAITS(PJRT_Extension_Base, next);
 // Changes include:
 // * Adding a new field to the PJRT_Api or argument structs
 // * Renaming a method or argument (doesn't affect ABI)
-#define PJRT_API_MINOR 106
+#define PJRT_API_MINOR 107
 
 // The plugin should set the major_version and minor_version of
 // PJRT_Api.pjrt_api_version to be the `PJRT_API_MAJOR` and `PJRT_API_MINOR` in
@@ -178,6 +178,28 @@ typedef enum {
   PJRT_Error_Code_UNAUTHENTICATED = 16
 } PJRT_Error_Code;
 
+typedef void (*PJRT_Error_PayloadVisitor)(const char* key, size_t key_size,
+                                          const char* value, size_t value_size,
+                                          void* user_arg);
+
+typedef struct PJRT_Error_FunctionTable {
+  size_t struct_size;
+  size_t instance_size;
+  PJRT_Extension_Base* extension_start;
+  void (*destroy)(PJRT_Error* error);
+  void (*message)(const PJRT_Error* error, const char** message,
+                  size_t* message_size);
+  PJRT_Error_Code (*get_code)(const PJRT_Error* error);
+  void (*for_each_payload)(const PJRT_Error* error,
+                           PJRT_Error_PayloadVisitor visitor, void* user_arg);
+} PJRT_Error_FunctionTable;
+PJRT_DEFINE_STRUCT_TRAITS(PJRT_Error_FunctionTable, for_each_payload);
+
+struct PJRT_Error {
+  const struct PJRT_Error_FunctionTable* vtable;
+};
+PJRT_DEFINE_STRUCT_TRAITS(PJRT_Error, vtable);
+
 struct PJRT_Error_GetCode_Args {
   size_t struct_size;
   PJRT_Extension_Base* extension_start;
@@ -187,10 +209,6 @@ struct PJRT_Error_GetCode_Args {
 PJRT_DEFINE_STRUCT_TRAITS(PJRT_Error_GetCode_Args, code);
 
 typedef PJRT_Error* PJRT_Error_GetCode(PJRT_Error_GetCode_Args* args);
-
-typedef void (*PJRT_Error_PayloadVisitor)(const char* key, size_t key_size,
-                                          const char* value, size_t value_size,
-                                          void* user_arg);
 
 struct PJRT_Error_ForEachPayload_Args {
   size_t struct_size;