+++ xla/pjrt/c/CHANGELOG.md
@@ -1,10 +1,17 @@
 # PJRT C API changelog
 
+## 0.107
+
+*   Added PJRT_Error_FunctionTable*.
+
+## 0.106
+
+*   Added `PJRT_Device_ClearMemoryStats` to `PJRT_Api` to allow resetting peak
+    memory tracking.
+
 ## 0.105
 
 *   Added PJRT_Memory_FunctionTable*.
-*   Added `PJRT_Device_ClearMemoryStats` to `PJRT_Api` to allow resetting peak
-    memory tracking.
 
 ## 0.104